Refactor: move dashboard types to their own types.ts
