+++ packages/admin-panel-server/src/viz-builder/dashboardVisualisation/combineVisualisation.ts
@@ -3,7 +3,8 @@
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
-import { DashboardItem, LegacyReport, Report, DashboardVizResource, DashboardViz } from '../types';
+import { LegacyReport, Report} from '../types';
+import { DashboardItem, DashboardViz, DashboardVizResource } from './types';
 
 const getData = (report: Report) => {
   const { config } = report;
