+++ packages/admin-panel-server/src/viz-builder/dashboardVisualisation/index.ts
@@ -6,3 +6,4 @@
 export { DashboardVisualisationExtractor } from './DashboardVisualisationExtractor';
 export { combineVisualisation } from './combineVisualisation';
 export * from './validators';
+export * from './types';
